Adiçao de chave estrangeira de user para cidade, criaçao de um novo metodo para cadastro de admin.
