+++ app/Http/Controllers/AdminMunicipioController.php
@@ -43,7 +43,7 @@ class AdminMunicipioController extends Controller
     //Metodo para listagem de estabelecimentos por cidade do Admin
     
     public function listEst($id) {
-        $admin = \App\AdminCidade::find($id);
+        $admin = \App\User::find($id);
         // $usuarios = \App\User::whereRaw('tipo like \'ADMIN\'')->select('id')->get();
         //$estabelecimentos = \App\Estabelecimento::whereIn("modalidade_id", $categorias)->get();
         //$estabelecimentos = \App\Estabelecimento::whereIn("iser_id", $usuarios)->get();
@@ -61,26 +61,5 @@ class AdminMunicipioController extends Controller
         }
         return dd($estabelecimentos);
         // return view("categoria.show")->with(['estabelecimentos' => $estabelecimentos]);
-    }
-
-    //Metodo para criar o usuário AdminCidade
-
-    public function AdminCreate(Request $request) {
-
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|string|max:255',
-            'email' => 'required|string|email|max:255|unique:users',
-            'password' => 'required|string|min:8|confirmed',
-        ]);
-
-        if(count($validator->errors()) > 0){
-            return redirect()->back()->withErrors($validator->errors())->withInput();
-        }
-
-        $dadosUsuarioAdmin = $request->only(["name", "email", "cidade_id"]);
-        $dadosUsuarioAdmin['password'] = Hash::make($request['password']);
-        $user = \App\AdminCidade::create($dadosUsuarioAdmin);
-
-        return $user;
     }
 }
